// src/data/data.interface.ts
export interface IData {
  bgImage: string;
  iconTrip: string;
  logo: string;
  imageRocket: string;
  title: string;
  subtitle: string;
  headerImage: string;
  footer: string;
  footerImage: string;
  hash: string;
  button: string;
  linkprev: string;
  linknext: string;
  easeView: string;
}